kraken: trying to reduce build time and memory
building data.cpp is take a very long time (more than 2minute) and a lot
of memory (around 4.5go). This seems to be caused by boost serialize.

In this PR I have started to move serialization function in .cpp to
limit the need of rebuilding it every time. I also moved each type in
it's own unit to reduce memory usage and prevent the rebuild of
everything. It should also speed up the build as each unit will be
smaller and build concurrently.

Also, I have added some forward declare in data.h, so a lot of files had to be updated
