+++ source/type/data.cpp
@@ -358,7 +358,6 @@ void Data::complete() {
     int admin, sort, autocomplete;
 
     build_grid_validity_pattern();
-    // build_associated_calendar(); read from database
 
     start = pt::microsec_clock::local_time();
     LOG4CPLUS_INFO(logger, "Building administrative regions");
@@ -387,94 +386,6 @@ void Data::complete() {
     LOG4CPLUS_INFO(logger, "\t Building admins: " << admin << "ms");
     LOG4CPLUS_INFO(logger, "\t Sorting data: " << sort << "ms");
     LOG4CPLUS_INFO(logger, "\t Building autocomplete " << autocomplete << "ms");
-}
-
-static ValidityPattern get_union_validity_pattern(const MetaVehicleJourney& meta_vj) {
-    ValidityPattern validity;
-    for (const auto& vj : meta_vj.get_base_vj()) {
-        if (validity.beginning_date.is_not_a_date()) {
-            validity.beginning_date = vj->base_validity_pattern()->beginning_date;
-        } else {
-            if (validity.beginning_date != vj->base_validity_pattern()->beginning_date) {
-                throw navitia::exception("the beginning date of the meta_vj are not all the same");
-            }
-        }
-        validity.days |= vj->base_validity_pattern()->days;
-    }
-    return validity;
-}
-
-void Data::build_associated_calendar() {
-    auto log = log4cplus::Logger::getInstance("kraken::type::Data");
-    std::multimap<ValidityPattern, AssociatedCalendar*> associated_vp;
-    size_t nb_not_matched_vj(0);
-    size_t nb_matched(0);
-    for (auto& meta_vj : this->pt_data->meta_vjs) {
-        // we check the theoric vj of a meta vj
-        // because we start from the postulate that the theoric VJs are the same VJ
-        // split because of dst (day saving time)
-        // because of that we try to match the calendar with the union of all theoric vj validity pattern
-        ValidityPattern meta_vj_validity_pattern = get_union_validity_pattern(*meta_vj);
-
-        // some check can be done on any theoric vj, we do them on the first
-        auto& first_vj = meta_vj->get_base_vj().front();
-        const auto& calendar_list = first_vj->route->line->calendar_list;
-        if (calendar_list.empty()) {
-            LOG4CPLUS_TRACE(log, "the line of the vj " << first_vj->uri << " is associated to no calendar");
-            nb_not_matched_vj++;
-            continue;
-        }
-
-        // we check if we already computed the associated val for this validity pattern
-        // since a validity pattern can be shared by many vj
-        auto it = associated_vp.find(meta_vj_validity_pattern);
-        if (it != associated_vp.end()) {
-            for (; it->first == meta_vj_validity_pattern; ++it) {
-                meta_vj->associated_calendars.insert({it->second->calendar->uri, it->second});
-            }
-            continue;
-        }
-
-        auto close_cal = find_matching_calendar(*this, meta_vj->uri, meta_vj_validity_pattern, calendar_list);
-
-        if (close_cal.empty()) {
-            LOG4CPLUS_TRACE(log, "the meta vj " << meta_vj->uri << " has been attached to no calendar");
-            nb_not_matched_vj++;
-            continue;
-        }
-        nb_matched++;
-
-        std::stringstream cal_uri;
-        for (auto cal_bit_set : close_cal) {
-            auto associated_calendar = new AssociatedCalendar();
-            pt_data->associated_calendars.push_back(associated_calendar);
-
-            associated_calendar->calendar = cal_bit_set.first;
-            // we need to create the associated exceptions
-            for (size_t i = 0; i < cal_bit_set.second.size(); ++i) {
-                if (!cal_bit_set.second[i]) {
-                    continue;  // cal_bit_set.second is the resulting differences, so 0 means no exception
-                }
-                ExceptionDate ex;
-                ex.date = meta_vj_validity_pattern.beginning_date + boost::gregorian::days(i);
-                // if the vj is active this day it's an addition, else a removal
-                ex.type = (meta_vj_validity_pattern.days[i] ? ExceptionDate::ExceptionType::add
-                                                            : ExceptionDate::ExceptionType::sub);
-                associated_calendar->exceptions.push_back(ex);
-            }
-
-            meta_vj->associated_calendars.insert({associated_calendar->calendar->uri, associated_calendar});
-            associated_vp.insert({meta_vj_validity_pattern, associated_calendar});
-            cal_uri << associated_calendar->calendar->uri << " ";
-        }
-
-        LOG4CPLUS_DEBUG(log, "the meta vj " << meta_vj->uri << " has been attached to " << cal_uri.str());
-    }
-
-    LOG4CPLUS_INFO(log, nb_matched << " vehicle journeys have been matched to at least one calendar");
-    if (nb_not_matched_vj) {
-        LOG4CPLUS_WARN(log, "no calendar found for " << nb_not_matched_vj << " vehicle journey");
-    }
 }
 
 /*
